+++ tybotRoute/tiledeskChatbotPlugs/directives/Directives.js
@@ -41,7 +41,7 @@ class Directives {
       name: action["_tdActionType"],//.type, //_tdActionType
       action: action
     }
-    delete directive.action["_tdActionType"];
+    // delete directive.action["_tdActionType"];
     console.log("Directive out:", directive);
     return directive;
   }